deep_learning/hw1/pyproject.toml
[tool.poetry]
name = "hw1"
version = "0.1.0"
description = ""
authors = ["toonnyy8 <toonnyy8@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch = "2.0.0"
numpy = "^1.25.2"
hydra-core = "^1.3.2"
einops = "^0.6.1"
matplotlib = "^3.7.2"
tqdm = "^4.65.2"
munch = "^4.0.0"
tensorboard = "^2.14.0"


[tool.poetry.group.dev.dependencies]
debugpy = "^1.6.7"
pyright = "^1.1.320"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
